% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance_matrix_utils.R
\name{add_flow_length}
\alias{add_flow_length}
\title{Add Length Column to Flow Data}
\usage{
add_flow_length(x)
}
\arguments{
\item{x}{sf object of flows (LINESTRING, projected CRS)}
}
\value{
sf object with an additional length_m column (od length in meters)
}
\description{
Also checks that 'origin' and 'destination' columns are present.
}
\examples{
flows <- sf::st_transform(flows_leeds, 3857)
flows <- add_flow_length(flows)
}
