% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance_matrix_utils.R
\name{add_xyuv}
\alias{add_xyuv}
\title{Add Start/End Coordinates & Flow IDs}
\usage{
add_xyuv(x)
}
\arguments{
\item{x}{sf object of flows}
}
\value{
tibble with x, y, u, v, flow_ID columns
}
\description{
Add Start/End Coordinates & Flow IDs
}
\examples{
flows <- sf::st_transform(flows_leeds, 3857)
flows <- add_flow_length(flows)
flows <- add_xyuv(flows)
}
