% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowcluster.R
\name{flowcluster}
\alias{flowcluster}
\title{Cluster origin-destination pairs (flows) with dbscan}
\usage{
flowcluster(x, eps, min_pts, dist_method = "euclidean")
}
\arguments{
\item{x}{A matrix of origin-destination pairs (flows) with two columns (TBC)}

\item{eps}{The maximum distance between two samples for one to be considered as in the neighborhood of the other. This is not a maximum bound on the distances of points within a cluster. This is the most important DBSCAN parameter to choose appropriately for your data set and distance function.}

\item{min_pts}{The number of samples in a neighborhood for a point to be considered as a core point. This includes the point itself.}

\item{dist_method}{The distance method to use. See \code{\link{dist}} for details.}
}
\value{
A vector with group assignments for each flow (TBC)
}
\description{
Cluster origin-destination pairs (flows) with dbscan
}
\examples{
\dontrun{
flowcluster(flowdata, eps = 0.1, min_pts = 5, dist_method = "euclidean")
}
}
