% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowcluster.R
\docType{data}
\name{flows_leeds}
\alias{flows_leeds}
\title{Example OD data}
\description{
A dataset containing origin-destination pairs (flows) for Leeds, UK.
The dataset contains origin, destination, count and geometry columns.
Count is the number of flows between origin and destination.
}
\note{
The schema data can be (re-)generated using code in the
\code{data-raw} directory.
}
\examples{
head(flows_leeds)
}
\keyword{datasets}
